Refactoring of KeywordCompleter
